게시판 CRUD

+++ src/main/resources/db/migration/V6__board.sql
@@ -1,6 +1,6 @@
 create table board
 (
-    id                bigint unsigned auto_increment comment 'ID' primary key,
+    id                int auto_increment comment 'ID' primary key,
     title             varchar(255)                               not null comment '제목',
     content           varchar(255)                               not null comment '내용',
     name              varchar(30)                                not null comment '이름',
